+++ CHANGELOG.md
@@ -9,7 +9,9 @@
     1. Dapat menyimpan report yang telah dibuat sehingga dapat ditinjau kembali. 
     2. Berbasis Query Builder sehingga lebih fleksibel dalam mengatur kriteria report yang dibuat.
     3. Dapat melihat SQL berdasarkan kriteria yang telah dibuat.
-- Peserta Event dapat diinput secara bulk menggunakan menu **Manage Peserta** pada event bagian peserta.
+    4. Duplikasi Nugie ataupun Rules Nugie yang telah dibuat.
+    5. Export hasil Nugie ke Excel.
+- Peserta Event dapat diinput secara bulk maupun file menggunakan menu **Manage Peserta** pada event bagian peserta.
 - Peserta Event sekarang yang akan atau sdh diinput dapat dilihat apakah sudah mengikuti diklat dengan kode kursus yang sama dan apakah mengikuti di event yang tanggalnya bertabrakan dengan event yang diikuti sekarang.
 - **Fungsi Export** untuk peserta event dan hasil dari Nugie.
 - **Panduan Pengguna** untuk semua fitur yang ada.